refactor(components): drop base component Ensure queue is updated when component args update by implementing update-queue modifer.
