+++ addon/components/file-upload.hbs
@@ -1,4 +1,14 @@
-<label for={{this.for}} class="file-upload">
+<label
+  for={{this.for}}
+  class="file-upload"
+  {{update-queue
+    this.queue
+    accept=@accept
+    multiple=@multiple
+    disabled=@disabled
+    onFileAdd=@onFileAdd
+  }}
+>
   <input
     id={{this.for}}
     type="file"
@@ -11,5 +21,4 @@
     {{on 'change' this.change}}
   />
 </label>
-
 {{yield this.queue}}
