+++ app/modifiers/update-queue.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-file-upload/modifiers/update-queue';
